requirements.txt
# python > 3.10.7
google-api-python-client
google-auth-httplib2
google-auth-oauthlib